+++ src/design/plone/contenttypes/behaviors/servizi_correlati.py
@@ -34,7 +34,6 @@ class IServiziCorrelati(model.Schema):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["Servizio"],
         },
     )