+++ src/design/plone/contenttypes/browser/utils/check_eventi.py
@@ -6,84 +6,40 @@ from openpyxl.styles import PatternFill
 from openpyxl.utils import get_column_letter
 from plone import api
 from Products.Five import BrowserView
-from zope.globalrequest import getRequest
-from zope.component import getMultiAdapter
-from plone.restapi.interfaces import ISerializeToJsonSummary
 
 import io
-
-from Acquisition import aq_inner
-from zope.component import getUtility
-from zope.intid.interfaces import IIntIds
-from zope.security import checkPermission
-from zc.relation.interfaces import ICatalog
 
 
 FLAG = '<i class="fa-solid fa-check"></i>'
 
 
-class CheckPersone(BrowserView):
+class CheckEventi(BrowserView):
     cds = None
 
     def is_anonymous(self):
         return api.user.is_anonymous()
 
-    def get_relations(self, obj, field):
-        return api.relation.get(source=obj, relationship=field, unrestricted=False)
+    def information_dict(self, evento):
+        prezzo = getattr(evento, "prezzo", "")
+        res = [x.get("text", "") for x in prezzo["blocks"].values()]
+        if not [x for x in res if x]:
+            prezzo = ""
 
-    def get_related_objects(self, obj, field):
-        """ """
-        items = []
-        relations = self.get_relations(obj, field)
-
-        for rel in relations:
-            rel_obj = rel.to_object
-            if rel_obj is not None:
-                summary = getMultiAdapter(
-                    (rel_obj, getRequest()), ISerializeToJsonSummary
-                )()
-                items.append(summary)
-        return sorted(items, key=lambda k: k["title"])
-
-    def back_references(self, source_object, attribute_name):
-
-        catalog = getUtility(ICatalog)
-        intids = getUtility(IIntIds)
-        result = []
-
-        for rel in catalog.findRelations(
-            dict(
-                to_id=intids.getId(aq_inner(source_object)),
-                from_attribute=attribute_name,
-            )
-        ):
-            obj = intids.queryObject(rel.from_id)
-            if obj is not None and checkPermission("zope2.View", obj):
-                result.append(obj)
-        return result
-
-    def information_dict(self, persona):
-        uo_refs = self.back_references(persona, "responsabile")
-        uo_refs.extend(self.back_references(persona, "persone_struttura"))
-
-        incarichi_persona = ""
-        if persona.incarichi_persona:
-
-            relations = self.get_related_objects(persona, "incarichi_persona")
-            if relations:
-                rel_data = relations[0]
-
-                if (
-                    rel_data["data_inizio_incarico"]
-                    and rel_data["title"].strip()
-                    and rel_data["tipologia_incarico"]
-                ):
-                    incarichi_persona = FLAG
+        luoghi_correlati = False
+        if getattr(evento, "luoghi_correlati", None):
+            luoghi_correlati = True
+        elif getattr(evento, "geolocation", None):
+            if getattr(evento.geolocation, "latitude", "") and getattr(
+                evento.geolocation, "longitude", ""
+            ):
+                luoghi_correlati = True
 
         return {
-            "incarichi_persona": incarichi_persona,
-            "contact_info": getattr(persona, "contact_info", None),
-            "organizzazione_riferimento": uo_refs,
+            "description": getattr(evento, "description", "").strip(),
+            "effective_date": getattr(evento, "effective_date", None),
+            "luoghi_correlati": luoghi_correlati,
+            "prezzo": prezzo,
+            "contact_info": getattr(evento, "contact_info", None),
         }
 
     def plone2volto(self, url):
@@ -95,47 +51,44 @@ class CheckPersone(BrowserView):
             return url.replace(portal_url, frontend_domain, 1)
         return url
 
-    def get_persone(self):
-
+    def get_eventi(self):
         if self.is_anonymous():
             return []
         pc = api.portal.get_tool("portal_catalog")
 
-        # show_inactive ha sempre avuto una gestione... particolare! aggiungo ai
-        # kw effectiveRange = DateTime() che è quello che fa Products.CMFPlone
-        # nel CatalogTool.py
         query = {
-            "portal_type": "Persona",
+            "portal_type": "Event",
             "review_state": "published",
+            "effectiveRange": DateTime(),
         }
-        brains = pc(query, **{"effectiveRange": DateTime()})
+        brains = pc(query)
         results = {}
         for brain in brains:
-            persona = brain.getObject()
+            evento = brain.getObject()
 
-            information_dict = self.information_dict(persona)
-
+            information_dict = self.information_dict(evento)
             if all(information_dict.values()):
                 continue
 
-            parent = persona.aq_inner.aq_parent
+            parent = evento.aq_inner.aq_parent
             if parent.title not in results:
                 results[parent.title] = {
                     "url": self.plone2volto(parent.absolute_url()),
                     "children": [],
                 }
-
             results[parent.title]["children"].append(
                 {
-                    "title": persona.title,
-                    "url": self.plone2volto(persona.absolute_url()),
+                    "title": evento.title,
+                    "description": information_dict.get("description") and FLAG or "",
+                    "url": self.plone2volto(evento.absolute_url()),
                     "data": {
-                        "organizzazione_riferimento": information_dict.get(
-                            "organizzazione_riferimento"
-                        )
+                        "effective_date": information_dict.get("effective_date")
                         and FLAG
                         or "",
-                        "incarichi_persona": information_dict.get("incarichi_persona"),
+                        "luoghi_correlati": information_dict.get("luoghi_correlati")
+                        and FLAG
+                        or "",
+                        "prezzo": information_dict.get("prezzo") and FLAG or "",
                         "contact_info": information_dict.get("contact_info")
                         and FLAG
                         or "",
@@ -150,15 +103,22 @@ class CheckPersone(BrowserView):
         return results
 
 
-class DownloadCheckPersone(CheckPersone):
+class DownloadCheckEventi(CheckEventi):
     CT = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
 
     def __call__(self):
-        HEADER = ["Titolo", "Incarichi", "Unità org. di riferimento", "Contatti"]
+        HEADER = [
+            "Titolo",
+            "Descrizione",
+            "Data",
+            "Luogo",
+            "Costo",
+            "Contatti",
+        ]
 
         EMPTY_ROW = [""] * 3
 
-        persone = self.get_persone()
+        eventi = self.get_eventi()
 
         workbook = Workbook()
         sheet = workbook.active
@@ -172,7 +132,7 @@ class DownloadCheckPersone(CheckPersone):
 
         section_row_height = int(14 * 1.5)
 
-        for category, category_data in persone.items():
+        for category, category_data in eventi.items():
             section_url = category_data["url"]
             section_title = category
             section_row = [section_title, "", ""]
@@ -205,15 +165,17 @@ class DownloadCheckPersone(CheckPersone):
                 column_letter = get_column_letter(col[0].column)
                 sheet.column_dimensions[column_letter].width = 35
 
-            for persona in category_data["children"]:
-                title_url = persona["url"]
-                dati_persona = [
-                    persona["title"],
-                    "X" if persona["data"]["incarichi_persona"] else "",
-                    "X" if persona["data"]["organizzazione_riferimento"] else "",
-                    "X" if persona["data"]["contact_info"] else "",
+            for evento in category_data["children"]:
+                title_url = evento["url"]
+                dati_evento = [
+                    evento["title"],
+                    "X" if evento["description"] else "",
+                    "X" if evento["data"]["effective_date"] else "",
+                    "X" if evento["data"]["luoghi_correlati"] else "",
+                    "X" if evento["data"]["prezzo"] else "",
+                    "X" if evento["data"]["contact_info"] else "",
                 ]
-                row = dati_persona
+                row = dati_evento
                 sheet.append(row)
 
                 title_cell = sheet.cell(row=sheet.max_row, column=1)
@@ -234,6 +196,6 @@ class DownloadCheckPersone(CheckPersone):
         self.request.RESPONSE.setHeader("Content-Type", self.CT)
         self.request.response.setHeader(
             "Content-Disposition",
-            "attachment; filename=check_persone.xlsx",
+            "attachment; filename=check_eventi.xlsx",
         )
         return data