+++ src/design/plone/contenttypes/events/common.py
@@ -3,6 +3,8 @@
 
 def onModify(context, event):
     for description in event.descriptions:
-        if "IBasic.title" in getattr(description, "attributes", []):
+        if "IBasic.title" in getattr(
+            description, "attributes", []
+        ) or "IDublinCore.title" in getattr(description, "attributes", []):
             for child in context.listFolderContents():
                 child.reindexObject(idxs=["parent"])